arch/x86/x86-locks: Provide a spin-lock implementation
This provides an x86 locking implementation of which causes a thread
trying to acquire the lock to simply wait in a loop /spin/ while
repeatedly checking if the lock is available. This is because some
synchronisation mechanism is required at the entry and exit of the
critical section to ensure exclusive use of certain shared hardware
resources. For example, as to not corrupt the serial UART telemetry.

Signed-off-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>